13/12/2021

+++ utils_2.c
@@ -6,10 +6,10 @@
 /*   By: Youssef <youssef.boughanmi.pro@gmail.com>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/07 19:51:47 by Youssef           #+#    #+#             */
-/*   Updated: 2021/12/11 09:50:02 by Youssef          ###   ########.fr       */
+/*   Updated: 2021/12/11 19:04:03 by Youssef          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include"my_printf.h"
+#include"ft_printf.h"
 
 int	fonc_unsigned(va_list argptr)
 {
@@ -24,9 +24,9 @@ int	fonc_unsigned(va_list argptr)
 int	fonc_hexa(va_list argptr, char type)
 {
 	int			add_char;
-	unsigned int	nbr;
+	unsigned long int	nbr;
 
-	nbr = va_arg(argptr, unsigned int);
+	nbr = va_arg(argptr, unsigned long int);
 	add_char = ft_unsitohexa(nbr, type);
 	return (add_char);
 }
